Add !define for prompts (#65)
Added Barret's old idea for letting users define gpt prompts then use
them as commands. It is used like so:

`!define prompt poem Write a poem about $args.`

In the future, if there is a need, I can add more variables like $name
or $time, etc. so that users can choose to inject more context to their
prompts or not. Originally I had it as $input (as the plurality of args
is confusing when you can't index it), but given that we're injecting
Message.args, I figured I'd keep the naming convention the same.

I don't want to do too fancy of functions yet (the Defn :func type). I
figure the implementation of that will likely involve the sandboxed WASM
Python that OpenAI uses for code interpreter, called over a Rust NIF.
That's long off.

Enjoy.