replace mogrify with image, fixes !motivate (#66)
No more need to install pango, Image ships with libvips (which has
pango). This also opens the door to more fun image commands.

We fixed some additional bugs for resolving (Telegram) `Attachment`s.
That being said, we also introduced some complexity. In the future,
let's simplify some.

In testing this, I noticed that Messages on my laptop has started
storing message text in a typestream (thankfully this hasn't happened on
the mini yet). But we'll need to add a typedstream reader in the future.